# artifacts/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KoriAPI",
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "paths": {
...
    "/v1/admin/overview": {
      "get": {
        "summary": "Admin overview",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/admin/logs": {
      "get": {
        "summary": "Structured logs",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/admin/audit": {
      "get": {
        "summary": "Audit trail",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/admin/devices": {
      "get": {
        "summary": "Device states",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/admin/provisioning-codes": {
      "post": {
        "summary": "Create provisioning code",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/admin/telemetry": {
      "get": {
        "summary": "Telemetry overview",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/admin/telemetry/enable-timescale": {
      "post": {
        "summary": "Enable Timescale",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/admin/jobs": {
      "get": {
        "summary": "List worker jobs",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/admin/quotas": {
      "get": {
        "summary": "List quota usage",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/admin/contracts": {
      "get": {
        "summary": "Contract manifest",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/admin/contracts/openapi.json": {
      "get": {
        "summary": "Generated OpenAPI document",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/admin/contracts/asyncapi.json": {
      "get": {
        "summary": "Generated AsyncAPI document",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "koriAuth": [
              "admin"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "koriAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-kori-session"
      }
    },
...
  }
}